+++ packages/client/src/components/main/Cart.tsx
@@ -6,6 +6,8 @@ import styled from 'styled-components'
 import { CartMenuType, SelectedMenuType } from 'type'
 import AmountController from './AmountController'
 import { IoRemoveOutline } from 'react-icons/io5'
+import Button from 'components/common/Button'
+import { CacheButtonList } from './CachePayment'
 
 interface CartProps {
   cartMenuList: CartMenuType[]
@@ -16,12 +18,6 @@ interface CartProps {
 interface CartMenuListProps {
   cartMenuList: CartMenuType[]
   updateCartMenuList: (newSelectedMenuList: CartMenuType[]) => void
-}
-
-interface ButtonProps {
-  value: string
-  onClick?: React.MouseEventHandler<HTMLButtonElement>
-  background?: string
 }
 
 function CartMenuList({ cartMenuList, updateCartMenuList }: CartMenuListProps) {
@@ -118,14 +114,6 @@ function Timer({ cartMenuList, deleteAllCartMenu }: TimerProps) {
   return <StyledTimer>{leftTime}초</StyledTimer>
 }
 
-function Button({ onClick, value, background }: ButtonProps) {
-  return (
-    <StyledButton onClick={onClick} background={background}>
-      {value}
-    </StyledButton>
-  )
-}
-
 export default function Cart({
   cartMenuList,
   updateCartMenuList,
@@ -155,17 +143,12 @@ export default function Cart({
           cartMenuList={cartMenuList}
           deleteAllCartMenu={deleteAllCartMenu}
         />
-        <Button
-          value="전체 취소"
-          onClick={onClickAllCancel}
-          background={theme.palette.warning}
-        />
-        <Button
-          value={`${getCartTotalAmount(
-            cartMenuList
-          ).toLocaleString()}원 결제하기`}
-          onClick={onClickPayButton}
-        />
+        <Button onClick={onClickAllCancel} background={theme.palette.warning}>
+          전체취소
+        </Button>
+        <Button onClick={onClickPayButton}>
+          {getCartTotalAmount(cartMenuList).toLocaleString()}원 결제하기
+        </Button>
       </div>
     </StyledCart>
   )
@@ -210,16 +193,4 @@ const StyledCartMenu = styled.li`
 const StyledTimer = styled.div`
   font-size: ${theme.font.lg};
   font-weight: 700;
-`
-
-const StyledButton = styled.button<{ background?: string }>`
-  width: 100%;
-  background-color: ${(props) => props.background || theme.palette.primary1};
-  padding: 1rem 3rem;
-  color: ${theme.palette.offWhite};
-  font-weight: 500;
-  font-size: ${theme.font.sm};
-
-  margin: 10px 0;
-  border-radius: 4px;
 `
